+++ changelog.md
@@ -36,6 +36,8 @@ All notable changes to this project will be documented in this file.
   ([#34](https://github.com/kotools/types/issues/34)).
 - Refactor binary operations of `kotools.types.number.StrictlyPositiveInt`
   ([#35](https://github.com/kotools/types/issues/35)).
+- Refactor binary operations of `kotools.types.number.NegativeInt`
+  ([#36](https://github.com/kotools/types/issues/36)).
 
 ## Version 1.2.0 - 2022/07/11